Fix coverage
